refactor: add class props descriptions
