+++ packages/oruga-next/src/components/slider/SliderThumb.vue
@@ -5,7 +5,7 @@ import OTooltip from "../tooltip/Tooltip.vue";
 
 import { isClient } from "@/utils/ssr";
 
-import type { SliderProps } from "./types";
+import type { SliderProps } from "./utils";
 import type { ClassBind } from "@/types";
 
 /**
